code/requirements.txt
numpy==2.2.2
pandas==2.2.3
python-dotenv==1.0.1
Requests==2.32.3
slack_sdk==3.34.0
snowflake_connector_python==3.13.2
SQLAlchemy==2.0.38
